Add test for string center and code refactor
